requirements.txt
# Core
pandas
sqlalchemy
pyodbc
python-dotenv

# Web scraping
beautifulsoup4
requests


# Validation & Testing
pydantic
pytest


# Orchestration
apache-airflow

# Dev
black
isort
